+++ api/card-service/src/main/java/nc/sgcb/labs/card/payment/web/CardPaymentResponse.java
@@ -7,7 +7,7 @@ public record CardPaymentResponse(
     @NotNull Long id,
     @NotNull Instant timestamp,
     @NotNull String currency,
-    @NotNull Long amount,
+    @NotNull Integer amount,
     @NotNull String cardNumber,
     @NotNull String destinationIban,
     boolean isAccepted) {
